tests: core/timer: relax system clock requirements
Calculate system clock resolution and use it as base for test timer
intervals. This allows timer-based tests to work on VMs and OSes with
crappy timekeeping.

+++ tools/mkscript.c
@@ -309,7 +309,7 @@ int main(int argc, const char *argv[])
 
         for (size_t j = 0; cname[j] != '\0'; j++)
         {
-            if (!isalnum(cname[j]))
+            if (!isalnum((int)cname[j]))
                 cname[j] = '_';
         }
